requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt[crypto]
django-filter